mental-health-rag/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
chromadb==0.4.22
pypdf==4.0.1
pdfplumber==0.11.0
pydantic==2.5.3
python-multipart==0.0.6
